🚚 Move : 유저 관련 기능 디렉토리 및 클래스명 정리
